+++ example-multi-activities/src/main/java/org/blockstack/android/DefaultConfig.kt
@@ -5,7 +5,7 @@ import org.blockstack.android.sdk.Scope
 val defaultConfig = java.net.URI("https://flamboyant-darwin-d11c17.netlify.com").run {
     org.blockstack.android.sdk.BlockstackConfig(
             this,
-            java.net.URI("${this}/redirect"),
-            java.net.URI("${this}/manifest.json"),
+            "/redirect",
+            "/manifest.json",
             kotlin.arrayOf(Scope.StoreWrite, Scope.Email))
 }